+++ flow-stream/src/main/scala/ch/jodersky/flow/stream/StreamWatcherException.scala
@@ -1,4 +1,0 @@
-package ch.jodersky.flow
-package stream
-
-class StreamWatcherException(message: String, cause: Throwable = null) extends RuntimeException(message, cause)