+++ stream/src/main/scala/akka/serial/stream/StreamWatcherException.scala
@@ -1,0 +1,4 @@
+package akka.serial
+package stream
+
+class StreamWatcherException(message: String, cause: Throwable = null) extends RuntimeException(message, cause)